+++ call-center/web-client/__mocks__/__utils__/createWebRTCCall.ts
