+++ call-center/web-client/src/components/Common/Common.test.tsx
@@ -3,9 +3,9 @@ import { rest } from 'msw';
 import { setupServer } from 'msw/node';
 import { render, waitFor, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import { BASE_URL } from '../configs/constants';
-import { IAgent } from '../interfaces/IAgent';
-import Common from '../components/Common';
+import { BASE_URL } from '../../configs/constants';
+import { IAgent } from '../../interfaces/IAgent';
+import Common from '.';
 
 jest.mock('@telnyx/webrtc');
